[webview] Remove "Cancel" from alert() dialogue. Contributes to JB#50018
According to W3Schools, a javascript alert() dialogue "displays an alert
box with a specified message and an OK button", in contrast with a
confirm() dialogue which also shows a Cancel button.

This change removes the "Cancel" text from the alert() dialogue (but
swiping in both directions is still possible and returns the same "OK"
value).