[sailfish-components-webview] Limit length of downloaded file name to 255 bytes. Fixes JB#52518
Default encoding of the file names is UTF-8. UTF-8 is a variable-width
character encoding. Some characters in this encoding take 2 or more bytes.
EXT4 has a maximum filename length of 255 bytes. This means that the
length of the filename should be limited taking into account the size of
the characters in UTF-8.