[sailfish-webview] Use gecko translation keys in dialogs. Contributes to JB#53307 OMP#JOLLA-219
Updates the default dialogs to use translation keys passed from the
gecko engine for localisation of messages presented to the user.

Updates the documentation with info about the translation keys for
custom dialog creation.

Adds a customisable Select custom dialog type for selecting multiple
values in a dialog with a message.