[sailfish-components-webview] Prefill prompt input field with default value. Contributes to JB#52387 OMP#JOLLA-257
When passed a default value to be prefilled into the prompt input
field, we should set the input field's text value to that value.

In order to prevent two-way-binding problem, we need to pass this
default value as a separate input property, rather than attempting
to bind it to the (output) value property.