[sailfish-webview] Act on private browsing when requesting permissions. JB#55925 OMP#JOLLA-445
A privateBrowsing flag is now passed with the "embed:permissions"
message when the browser wants to request a permission of the user (a
location permission specifically).

This change disables the option to store the permission when private
browsing is enabled, to avoid leaking location data.

In addition, the info is now passed using the new
LocationPopupInterface.privateBrowsing flag added to the brower popups
API.